Rename messageActions.ts -> ChannelActions.ts
