+++ packages/shared/src/index.ts
@@ -8,7 +8,7 @@ export type {
   WebAppAction,
   WebAppCommand,
 } from './types/native';
-export { getPostActions, ChannelAction } from './types/messageActions';
+export { getPostActions, ChannelAction } from './types/ChannelActions';
 export {
   PostCollectionConfiguration,
   usePostCollectionConfigurationFromChannel,
